Library API

The api is categorized by the underlying basis being used

Chebyshev

DMSuite.chebdif — Function

chebdif(N, M)

Computes the differentiation matrices D1, D2, ..., DM on Chebyshev nodes.

Arguments

N: size of differentiation matrix.

M: number of derivatives required [integer]. Note: 0 < M <= N-1.

Outputs

x: the Chebyshev nodes.

DM: DM[1:N,1:N,ell] contains ell-th derivative matrix, ell=1,...,M.

Details

The code implements two strategies for enhanced accuracy suggested by W. Don & S. Solomonoff in SIAM J. Sci. Comp. Vol. 6, pp. 1253–1268 (1994). The two strategies are (a) the use of trigonometric identities to avoid the computation of differences x[k]-x[j] & (b) the use of the "flipping trick" which is necessary since sin(t) can be computed to high relative precision when t is small whereas sin(pi-t) cannot.

DMSuite.chebdifft — Function

chebdifft(f,M)

Computes the M'th approximate Chebyshev derivatives of the data vector f, using the FFT.

Arguments

f: vector of length N containing function values at the Chebyshev points x[k] = cos((k-1)*pi/(N-1)), k = 1...N.

M: derivative required [positive integer]

Outputs

Dmf: vector containing approximate M'th derivative

Details

A Fast Fourier Transform is used compute the Chebyshev cofficients of the data vector. A recursion formula is used to compute the Chebyshev coefficients for each derivative. A FFT is then used again to compute the derivatives in physical space.

DMSuite.chebint — Function

chebint(fk, x)

Evaluates the polynomial interpolant of the data points (xk[j], fk[j]), where xk[j] are the Chebyshev nodes, at the points x. Requires two or more data points.

Arguments

fk: vector of y-coordinates of data; at Chebyshev points xk[j] = cos((j-1)*pi/(N-1)), j = 1...N.

x: vector of x-values where polynomial interpolant is to be evaluated.

Outputs

p: vector of interpolated values.

Details

The code implements the barycentric formula; see page 252 in P. Henrici; Essentials of Numerical Analysis; Wiley; 1982.

DMSuite.cheb2bc — Function

cheb2bc(N,g)

Computes the first and second derivative matrices, as well as boundary condition functions, for 2 point boundary conditions:

a1 u[1] + b1 u'[1] = c1

aN u[-1] + bN u'[-1] = cN

Arguments

N: number of Chebyshev points in [-1,1]

g: boundary condition matrix = [a1 b1 c1; aN bN cN]

Outputs

xt: Chebyshev points corresponding to rows and columns of D1t & D2t

D1t: 1st derivative matrix incorporating bc

D2t: 2nd derivative matrix incorporating bc

phip: 1st & 2nd derivative of bc function at x=1 (array with 2 columns)

phim: 1st & 2nd derivative of bc function at x=-1 (array with 2 columns)

DMSuite.cheb4c — Function

cheb4c(N)

Computes the fourth derivative matrix on Chebyshev interior points, incorporating the clamped boundary conditions:

u[1]=u"[1]=u[-1]=u"[-1]=0.

Arguments

N: N-2 = order of differentiation matrix. (The interpolant has degree N+1.)

Outputs

x: interior Chebyshev points [vector of length N-2]

D4: fourth derivative matrix [size (N-2)x(N-2)]

Details

The code implements two strategies for enhanced accuracy suggested by W. Don & S. Solomonoff in SIAM J. Sci. Comp. Vol. 6, pp. 1253–1268 (1994). The two strategies are (a) the use of trigonometric identities to avoid the computation of differences x[k]-x[j] & (b) the use of the "flipping trick" which is necessary since sin(t) can be computed to high relative precision when t is small whereas sin(pi-t) cannot.

DMSuite.cheb1roots — Function

cheb1roots(N)

Computes the roots of the Chebyshev polynomial of the first kind of degree N. AKA, Chebyshev-Gauss (CG) Points, or Chebyshev points of the first kind.

DMSuite.cheb1extrema — Function

cheb1extrema(N)

Computes the extrama of the Chebyshev polynomial of the first kind of degree N. AKA, Chebyshev-Gauss-Lobatto (CGL) Points, or Chebyshev points of the second kind.

Hermite, Laguerre, and Legendre

DMSuite.herdif — Function

herdif(N, M, b)

Computes the differentiation matrices D1, D2, ..., DM on Hermite points.

Arguments

N: number of points, i.e., order of differentiation matrices [integer].

M: number of derivatives required [integer]. Note that M must satisfy, 0 < M < N-1.

b: scaling parameter [real, positive].

Outputs

x: vector of nodes (zeros of Hermite polynomial of degree N scaled by the parameter b.)

DM: DM[1:N,1:N,ell] contains ell-th derivative matrix, ell=1,..,M.

DMSuite.lagdif — Function

lagdif(N, M, b)

Computes the differentiation matrices D1, D2, ..., DM on Laguerre points.

Arguments

N: number of points, i.e., order of differentiation matrices [integer].

M: number of derivatives required [integer]. Note that M must satisfy, 0 < M < N-1.

b: scaling parameter [real, positive].

Outputs

x: vector of nodes (zeros of Laguerre polynomial of degree N-1 plus x = 0), all scaled by the parameter b.

DM: DM[1:N,1:N,ell] contains ell-th derivative matrix, ell=1,...,M.

DMSuite.herroots — Function

herroots(N)

Computes the roots of the Hermite polynomial of degree N.

DMSuite.lagroots — Function

lagroots(N)

Computes the roots of the Laguerre polynomial of degree N.

DMSuite.legroots — Function

legroots(N)

Computes the roots of the Legendre polynomial of degree N.

Generic Polynomial

DMSuite.poldif — Function

poldif(x, alpha, B)

Computes the differentiation matrices D1, D2, ..., DM on arbitrary nodes, x, with associated weights, alpha.

Arguments

x: vector of N distinct nodes.

alpha: vector of weight values alpha(x), evaluated at x = x[k].

B: matrix of size M x N; where M is the highest derivative required. It should contain the quantities

B[ell,j] = beta(ell,j) = (ell-th derivative of alpha(x[j]))/alpha(x[j]). Note that M must satisfy, 0 < M < N-1.

Outputs

DM: DM[1:N,1:N,ell] contains ell-th derivative matrix, ell=1,...,M.

poldif(x, M)

Computes the differentiation matrices D1, D2, ..., DM on arbitrary nodes, x, with unit weights.

Arguments

x: vector of N distinct nodes.

M: the number of derivatives required [integer]. Note that M must satisfy, 0 < M < N-1.

Outputs

DM: DM[1:N,1:N,ell] contains ell-th derivative matrix, ell=1,...,M.

DMSuite.polint — Function

polint(xk, fk, x, alpxk, alpx)

Computes the polynomial interpolant of the data [xk, fk] under the provided weights, alpxk and alpx.

Arguments

xk: vector of x-coordinates of data [assumed distinct].

fk: vector of y-coordinates of data.

x: vector of x-values where polynomial interpolant is to be evaluated.

alpxk: vector of weight values sampled at the points xk.

alpx: vector of weight values sampled at the points x.

Outputs

p: vector of interpolated values.

Details

The code implements the barycentric formula; see page 252 in P. Henrici; Essentials of Numerical Analysis; Wiley; 1982.

Except for certain nice node distributions polynomial interpolation of high-degree is an ill-conditioned problem. This code does not test for conditioning so use with care.

polint(xk, fk, x)

Evaluates the polynomial interpolant of the data [xk, fk], at the points x. Requires two or more data points.

Arguments

xk: vector of x-coordinates of data [assumed distinct].

fk: vector of y-coordinates of data.

x: vector of x-values where polynomial interpolant is to be evaluated.

Outputs

p: Vector of interpolated values.

Details

Except for certain nice node distributions polynomial interpolation of high-degree is an ill-conditioned problem. This code does not test for conditioning so use with care.

Fourier

DMSuite.fourdif — Function

fourdif(N,M)

Computes the M'th derivative Fourier spectral differentiation matrix on grid with N equispaced points in [0,2pi)

Arguments

N: size of differentiation matrix.

M: derivative required [non-negative integer]

Outputs

x: equispaced points 0, 2pi/N, 4pi/N, ... , (N-1)2pi/N

DM: M'th order differentiation matrix

Details

Explicit formulas are used to compute the matrices for M=1 & 2. A discrete Fouier approach is employed for M>2. The program computes the first column and first row & then uses the toeplitz command to create the matrix. For M=1 & 2 the code implements a "flipping trick" to improve accuracy suggested by W. Don & A. Solomonoff in SIAM J. Sci. Comp. Vol. 6, pp. 1253–1268 (1994). The flipping trick is necesary since sin(t) can be computed to high relative precision when t is small whereas sin(pi-t) cannot.

DMSuite.fourdifft — Function

fourdifft(f,M)

Computes the M-th derivative of the function f[x] using the Fourier differentiation process. The function is assumed to be 2pi-periodic & the input data values f should correspond to samples of the function at N equispaced points on [0, 2pi). The Fast Fourier Transform is used.

Arguments

f: vector of samples of f[x] at x = 0, 2pi/N, 4pi/N, ... , (N-1)2pi/N

M: derivative required [non-negative integer]

Outputs

Dmf: M-th derivative of f

DMSuite.fourint — Function

fourint(fk, x)

Evaluates the trigonometric interpolant of the data (xk[j], fk[j]), where xk are equidistant nodes, at the points x. Requires two or more data points

Arguments

fk: vector of y-coordinates of data, at equidistant points xk[j] = (j-1)2pi/N, j = 1...N.

x: vector of x-values where interpolant is to be evaluated.

Outputs

t: vector of interpolated values.

Details

The code implements the barycentric formula; see page 46 in P. Henrici; Applied & Computational Complex Analysis III; Wiley; 1986.

Sinc

DMSuite.sincdif — Function

sincdif(N, M, h)

Computes sinc the differentiation matrices D1, D2, ..., DM on equidistant points.

Arguments

N: number of points; i.e.; order of differentiation matrix.

M: number of derivatives required [integer]. Note that M must satisfy, 0 < M < N-1.

h: step-size (real, positive).

Outputs

x: vector of nodes.

DM: DM[1:N,1:N,ell] contains ell-th derivative matrix, ell=1,...,M.

DMSuite.sincdifft — Function

sincdifft(f, M, h)

Computes the m-th derivative of the function f[x] using the sinc differentiation process. The function is assumed to be defined on the entire real line & the input values correspond to samples of the function at N equispaced points symmetric with respect to the origin.

Arguments

f: vector of samples of f[x] at h*[-(N-1)/2:(N-1)/2]

M: number of derivatives required [integer]. Note that M must satisfy, 0 < M < N-1.

h: step-size (real, positive).

Outputs

Dmf: m-th derivative of f